test(gateway): move archived-confirmation case to the outcomes suite
The authority-and-recovery suite sits at the oxlint max-lines ceiling, so
the new archived-session case belongs with the personal publication
outcomes tests that share the same fixture.
